refactor: rename HandleEmbedPagination to HandlePagePagination and update imports across multiple files
