+++ Source/Commands/Miscellaneous/Duty/Duty Types/Subcmds/View.ts
@@ -2,7 +2,7 @@ import { Guilds } from "@Typings/Utilities/Database.js";
 import { InfoEmbed } from "@Utilities/Classes/ExtraEmbeds.js";
 import { SlashCommandSubcommandBuilder, EmbedBuilder, Colors, MessageFlags } from "discord.js";
 
-import HandlePagePagination from "@Utilities/Other/HandleEmbedPagination.js";
+import HandlePagePagination from "@Utilities/Other/HandlePagePagination.js";
 import GetShiftTypes from "@Utilities/Database/GetShiftTypes.js";
 import Dedent from "dedent";
 import Util from "node:util";
